refactor(file): 파일/폴더 구조 리팩토링
