+++ src/app/(global)/detail/[id]/page.tsx
@@ -1,12 +1,13 @@
 interface DetailPageProps {
-  params: { id: string };
+  params: Promise<{ id: string }>;
 }
 
-export default function DetailPage({ params }: DetailPageProps) {
+export default async function DetailPage({ params }: DetailPageProps) {
+  const { id } = await params;
   return (
     <main>
       <h1>체험 상세 페이지</h1>
-      <p>체험 ID: {params.id}</p>
+      <p>체험 ID: {id}</p>
     </main>
   );
 }
