+++ src/components/pages/detail/Calendar.tsx
@@ -5,7 +5,7 @@ import ArrowLeft from '@/public/images/icons/ArrowLeft.svg';
 import ArrowRight from '@/public/images/icons/ArrowRight.svg';
 import { useCalendar } from '@/src/hooks/useCalendar';
 import { useCallback } from 'react';
-import CalendarDay from '@/src/components/CalendarDay';
+import CalendarDay from '@/src/components/primitives/CalendarDay';
 
 export default function Calendar({
   dateCallback,
