+++ src/components/primitives/global/Footer.tsx
